Replace db array default values with real literals
The default values for the database transfer functions were incorrectly
assiged as integer literals, recast to types using real32/64 but
actually corresponding to whatever integer kind equals real32/64.

We now simply assign it a literal value of -1. and rely on the compiler
to handle the recasting.

Although none of these functions were intended to be used, and -1 would
probably be eventually cast into an appropriate real type, it is better
to get this correct.

Thanks to Keith Lindsay for suggesting this change.