Added missing values for vector u2_bg(:) within set_viscous_ML
Vector u2_bg(:) was introduced in commit 6c44c5f38 (Fix for using
tidal amplitude in determining the BBL thickness, 2024-04-04). It
replaces a scalar in set_viscous_ML, but was not defined, causing
crashes. This commit sets the values of u2_bg(:) to match those given
given in set_viscous_BBL.