GitLab CI: Direct build directory to $HOME
This patch directs the compilation of GitLab regression test models to
the $HOME directory.

This is motivated by incredibly slow file access times on GFDL's GPFS
file system (F5), with the Nvidia builds taking a prohibitively long
time (around 4 hours).  Compiling in $HOME reduces Nvidia times to 15
minutes, with similar reduction times in the other compilation.

This is achieved with only minimal changes to the build system.  Instead
of creating `build/`, we create a dedicated directory in
`$HOME/ci/$CI_PROJECT_ID/build` and a symbolic link to this directory in
the GPFS directory.  To support these builds, additional symlinks to
`src/` and `ocean_only/` (for static builds) are also created, from
$HOME and back into GPFS.

Once the job is completed, the temporary space in $HOME is deleted.

This patch should not be considered a permanent solution, it is only
meant to get us through a difficult period of testing.  Hopefully we can
revert this change in the near future.