Replace kind=8 with iso_fortran_env kinds
This patch replaces references to kind=8 with the corresponding Fortran
kind identifiers as defined in iso_fortran_env.

Although these are widely equal to 8, there are a number of compilers
(notably NAG, also others) which set it to a different value.

This improves the portability of the code (albeit in a rather minor
way).