Correct 10 do-loop index cases
  Corrected the case of the do-loop index variable in 10 places in 6 files,
initialized two whole diagnostic arrays to 0 instead of having shared loops
initializing u- and v-point arrays, and shortened two loops initializing
pressures to the size that is actually used in the corresponding equation of
state calls.  With these changes, there are no i- or j- loops with extents that
are inconsistent with the stagger-based index case convention used throughout
the MOM6 code.  All answers are bitwise identical in all cases.